Split account page test cases in two files to run faster
