+++ .circleci/config.yml
@@ -28,7 +28,7 @@ jobs:
   cypress/run:
     docker:
     - image: cypress/base:8
-    parallelism: 4
+    parallelism: 24
     steps:
     - run:
         command: echo "Assuming dependencies were installed using cypress/install
