+++ test_filenames.txt
@@ -21,3 +21,4 @@ tests/e2e/specs/settings.js
 tests/e2e/specs/transactionsPage.js
 tests/e2e/specs/utilitiesPage.js
 tests/e2e/specs/signAndVerifyMessage.js
+tests/e2e/specs/accountPage2.js
